requirements.txt
-----VERSION below-----
numpy==1.17.4
nibabel==2.5.1
nilearn==0.6.2
tables==3.6.1
tensorflow-gpu=1.14.0
Keras==2.0.8
SimpleITK==2.0.0
h5py==2.9.0
pip install git+https://www.github.com/keras-team/keras-contrib.git
